+++ App/components/PlayerGameDetailScreen/index.js
@@ -1,0 +1,3 @@
+import PlayerGameDetailScreen from './PlayerGameDetailScreen'
+
+export default PlayerGameDetailScreen