+++ App/components/ScoresScreen/GameCell.js
@@ -2,7 +2,20 @@ import React, { Component } from 'react'
 import { Text, View, StyleSheet, TouchableOpacity, Image } from 'react-native'
 import { connect } from 'react-redux'
 import { selectGame } from '../../actions/actions'
+import Circle from '../common/Circle'
 import TeamMap from '../../utils/TeamMap'
+
+const TeamCircle = ({ teamName, teamColor }) => (
+  <Circle color={teamColor} height={60} width={60} borderRadius={120}>
+    <View style={styles.defaultCenteredView}>
+      <Circle color='#111111' height={55} width={55} borderRadius={110}>
+        <View style={styles.defaultCenteredView}>
+          <Text style={[styles.text, { fontSize: 15 }]}> {teamName} </Text>
+        </View>
+      </Circle>
+    </View>
+  </Circle>
+)
 
 const TeamSection = ({style, team, teamColor, logo, flip}) => (
   <View style={[styles.team, style]}>
@@ -18,7 +31,10 @@ const TeamSection = ({style, team, teamColor, logo, flip}) => (
           style={styles.logo}
           source={logo}
         /> */}
-        <Text style={[styles.text, { fontSize: 22, color: teamColor }]}> {team.triCode} </Text>
+        <TeamCircle
+          teamColor={teamColor}
+          teamName={team.triCode}
+        />
       </View>
     </View>
     {
@@ -109,9 +125,7 @@ class GameCell extends Component<Props> {
             logo={awayTeamLogo}
             flip={false}
           />
-          {
-            this._renderGameStatus()
-          }
+          { this._renderGameStatus() }
           <TeamSection
             style={styles.homeTeam}
             team={hTeam}
@@ -128,7 +142,8 @@ class GameCell extends Component<Props> {
 const styles = StyleSheet.create({
   text: {
     textAlign: 'center',
-    color: '#D3D3D3'
+    color: '#D3D3D3',
+    fontFamily: 'Rubik-Light'
   },
   defaultCenteredView: {
     flex: 1,
@@ -151,7 +166,6 @@ const styles = StyleSheet.create({
     alignItems: 'center',
     height: 120,
     backgroundColor: '#141414',
-    // backgroundColor: '#212121',
     flexDirection: 'row',
     borderBottomLeftRadius: 12,
     borderBottomRightRadius: 12,
@@ -169,7 +183,6 @@ const styles = StyleSheet.create({
     marginLeft: 10
   },
   gameInfo: {
-    flex: 1,
     justifyContent: 'center',
     alignItems: 'center',
     flexDirection: 'column'