+++ App/components/ScoresScreen/Header.js
@@ -35,6 +35,9 @@ class Header extends Component<Props> {
     else if (type === 'previous') {
       date = new Date(date.getTime() - (24 * 60 * 60 * 1000)).toLocaleDateString()
     }
+    else if (type === 'today') {
+      date = this.props.currentDate
+    }
 
     this.props.changeDateToNewDate(date)
   }
@@ -50,7 +53,7 @@ class Header extends Component<Props> {
     return (
       <TouchableOpacity onPress={() => { this.handleDateChange(handler) }}>
         <View>
-          <Text style={{fontSize: 36}}> { direction } </Text>
+          <Text style={{fontSize: 36, fontFamily: 'Rubik-Light'}}> { direction } </Text>
         </View>
       </TouchableOpacity>
     )
@@ -62,13 +65,13 @@ class Header extends Component<Props> {
         <View style={styles.statusBar} />
         <View style={styles.dateCointainer}>
           { this._renderArrow('<') }
-          <TouchableOpacity style={styles.dateCointainerMiddle} onPress={() => { this.handleDateChange('other') }}>
+          <TouchableOpacity style={styles.dateCointainerMiddle} onPress={() => { this.handleDateChange('today') }}>
             <View style={{flex: 1}} >
               <View style={styles.defaultCenteredView}>
-                <Text style={{fontSize: 18, color: '#FFFFFF'}}> { this.formatDate() } </Text>
+                <Text style={{fontSize: 18, color: '#FFFFFF', fontFamily: 'Rubik-Light'}}> { this.formatDate() } </Text>
               </View>
               <View style={styles.defaultCenteredView}>
-                <Text style={{fontSize: 16, color: '#FFFFFF'}}> { this.getNumberOfGamesText() } </Text>
+                <Text style={{fontSize: 16, color: '#FFFFFF', fontFamily: 'Rubik-Light'}}> { this.getNumberOfGamesText() } </Text>
               </View>
             </View>
           </TouchableOpacity>
@@ -104,7 +107,8 @@ const styles = StyleSheet.create({
 
 function mapStateToProps(state) {
   return {
-    date: state.date.date
+    date: state.date.date,
+    currentDate: state.date.currentDate
   }
 }